% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Complete Descriptive Statistics of Every Column in a Dataframe}
\usage{
descriptive_stats(dataframe, csv_file = FALSE, alternative.class_df = NULL)
}
\arguments{
\item{dataframe}{A dataframe}

\item{csv_file}{Logical, save results in a .csv file}

\item{alternative.class_df}{Dataframe used to assign a class to every column. Must be like get_correct_class output}
}
\value{
a dataframe or a .csv file with the complete descriptive statistics of the initial dataframe
}
\description{
Complete Descriptive Statistics of Every Column in a Dataframe
}
\examples{
sample_data.descriptive = descriptive_stats(sample_data)
}
