% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correct_class.R
\name{get_correct_class}
\alias{get_correct_class}
\title{Obtain Class Indication For Every Column Of A Dataframe}
\usage{
get_correct_class(dataframe, unq.max.perc = 0.15)
}
\arguments{
\item{dataframe}{A dataframe}

\item{unq.max.perc}{cutoff frequency of the most common variable to distinguish CAT from NUM}
}
\value{
a dataframe with a column of variables and a column of data types
}
\description{
Obtain Class Indication For Every Column Of A Dataframe
}
\examples{
sample_data.class = get_correct_class(sample_data)
}
