% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data}
\format{
\subsection{\code{sample_data}}{

A tibble with 50 rows and 9 columns:
\describe{
\item{ID}{1 value x column}
\item{sex}{2 categories}
\item{age}{integer, multiple values}
\item{BMI}{numeric, with decimals}
\item{grade}{integer, 4 different values}
\item{treated}{TRUE/FALSE column}
\item{event_1}{column with some NAs}
\item{event_2}{column filled with NA}
\item{date_event}{random dates}
}
}
}
\source{
random made up data
}
\usage{
sample_data
}
\description{
A random set of data, all values are made up
}
\keyword{datasets}
