% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_descriptive_stats.R
\name{subset_descriptive_stats}
\alias{subset_descriptive_stats}
\title{Descriptive Statistics for Subsets of a Dataframe}
\usage{
subset_descriptive_stats(dataframe, group_var, csv_file = FALSE)
}
\arguments{
\item{dataframe}{A dataframe}

\item{group_var}{The name of a column in the dataframe}

\item{csv_file}{Logical, save results in a .csv file}
}
\value{
a dataframe or a .csv file with descriptive statistics for each subset based on the group_var.
}
\description{
Descriptive Statistics for Subsets of a Dataframe
}
\examples{
sample_data.subset_descriptive = subset_descriptive_stats(sample_data, "sex")
}
